requirements.txt
pymongo
python-dotenv
pandas
numpy
sentence-transformers
faiss-cpu
pdfplumber
camelot-py[cv]
python-docx
seaborn
matplotlib
streamlit
transformers
datasets
peft
accelerate
scikit-learn
watchdog
